+++ App/Views/Map/index.view.php
@@ -16,22 +16,27 @@ use App\Models\Kategorymap;
 <body>
 <div class="row">
     <div class="column side">
-
+        <?php foreach ($data['data'] as $map) { ?>
+            <p></p>
+            <a type="button" href="#<?=$map->getId()?>" class="btn btn-dark btn-lg row"><?=$map->getNazov()?></a>
+        <?php } ?>
     </div>
     <div class="column middle">
         <div class="container-fluid">
             <h1 class="NadpisSekcia">Mapy každého druhu</h1>
-            <div class="row">
-                <div class="col">
-                    <a href="?c=map&a=create" class="btn btn-success">Pridať novú mapu</a>
+            <?php if ($auth->isLogged()) { ?>
+                <div class="row">
+                    <div class="col">
+                        <a href="?c=map&a=create" class="btn btn-success">Pridať novú mapu</a>
+                    </div>
                 </div>
-            </div>
+            <?php } ?>
             <div class="row">
                 <?php foreach ($data['data'] as $map) { ?>
                     <div class="col-xl-4 col-md-4 col-sm-6">
                         <div class="card my-3">
                             <div class="card-body">
-                                <h5 class="card-title">
+                                <h5 id="<?= $map->getId() ?>" class="card-title">
                                     <?= $map->getNazov() ?>
                                 </h5>
 
@@ -43,8 +48,12 @@ use App\Models\Kategorymap;
                                 <p></p>
                                 <p align="center">
                                     <a href="?c=map&a=open&id=<?= $map->getId() ?>" class="btn btn-success">Otvoriť</a>
-                                    <a href="?c=map&a=edit&id=<?= $map->getId() ?>" class="btn btn-warning">Upraviť</a>
-                                    <a href="?c=map&a=delete&id=<?= $map->getId() ?>" class="btn btn-danger">Zmazať</a>
+                                    <?php if ($auth->isLogged()) { ?>
+                                        <?php if ($map->getAutor() == $auth->getLoggedUserId()) { ?>
+                                            <a href="?c=map&a=edit&id=<?= $map->getId() ?>" class="btn btn-warning">Upraviť</a>
+                                            <a href="?c=map&a=delete&id=<?= $map->getId() ?>" class="btn btn-danger">Zmazať</a>
+                                        <?php } ?>
+                                    <?php } ?>
                                 </p>
                             </div>
                         </div>
